:recycle: api: Move shared/infrastructure/mailer to shared/mail/infrastructure
