+++ api/src/shared/mail/domain/models/EmailingAttempt.js
