+++ api/src/shared/mail/domain/models/MailingProviderInvalidEmailError.js
