+++ api/tests/acceptance/application/session/session-controller-import-certification-candidates-from-attendance-sheet_reports_categorization_test.js
@@ -1,6 +1,6 @@
 import { databaseBuilder, expect, generateValidRequestAuthorizationHeader, knex, sinon } from '../../../test-helper.js';
 import { createServer } from '../../../../server.js';
-import { clearResolveMx, setResolveMx } from '../../../../lib/infrastructure/mail-check.js';
+import { clearResolveMx, setResolveMx } from '../../../../src/shared/mail/infrastructure/smtp/mail-check.js';
 import fs from 'fs';
 import * as url from 'url';
