+++ api/tests/shared/mail/domain/services/MailingProvider_test.js
@@ -1,5 +1,5 @@
 import { expect } from '../../../../test-helper.js';
-import { MailingProvider } from '../../../../../src/shared/infrastructure/mailers/MailingProvider.js';
+import { MailingProvider } from '../../../../../src/shared/mail/infrastructure/smtp/MailingProvider.js';
 
 describe('Unit | Infrastructure | Mailers | MailingProvider', function () {
   describe('#sendEmail', function () {
